implemented piece placement UI, reorganized files
