--- Assets/MoveInput.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b948523af4c873949b458cc0123775cd
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
