+++ Assets/Scenes/LeoScene.unity
@@ -113,6 +113,53 @@ NavMeshSettings:
     debug:
       m_Flags: 0
   m_NavMeshData: {fileID: 0}
+--- !u!1001 &67106987
+Prefab:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    m_TransformParent: {fileID: 0}
+    m_Modifications:
+    - target: {fileID: 4136581459228970, guid: b5d94b40f41174652b848ee15d2a7adf, type: 2}
+      propertyPath: m_LocalPosition.x
+      value: 100
+      objectReference: {fileID: 0}
+    - target: {fileID: 4136581459228970, guid: b5d94b40f41174652b848ee15d2a7adf, type: 2}
+      propertyPath: m_LocalPosition.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 4136581459228970, guid: b5d94b40f41174652b848ee15d2a7adf, type: 2}
+      propertyPath: m_LocalPosition.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 4136581459228970, guid: b5d94b40f41174652b848ee15d2a7adf, type: 2}
+      propertyPath: m_LocalRotation.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 4136581459228970, guid: b5d94b40f41174652b848ee15d2a7adf, type: 2}
+      propertyPath: m_LocalRotation.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 4136581459228970, guid: b5d94b40f41174652b848ee15d2a7adf, type: 2}
+      propertyPath: m_LocalRotation.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 4136581459228970, guid: b5d94b40f41174652b848ee15d2a7adf, type: 2}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 4136581459228970, guid: b5d94b40f41174652b848ee15d2a7adf, type: 2}
+      propertyPath: m_RootOrder
+      value: 8
+      objectReference: {fileID: 0}
+    - target: {fileID: 114739047883504446, guid: b5d94b40f41174652b848ee15d2a7adf,
+        type: 2}
+      propertyPath: board
+      value: 
+      objectReference: {fileID: 597238866}
+    m_RemovedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: b5d94b40f41174652b848ee15d2a7adf, type: 2}
+  m_IsPrefabAsset: 0
 --- !u!1 &69620649
 GameObject:
   m_ObjectHideFlags: 0
@@ -265,7 +312,7 @@ Transform:
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children: []
   m_Father: {fileID: 959419239}
-  m_RootOrder: 2
+  m_RootOrder: 3
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!114 &125722436
 MonoBehaviour:
@@ -278,9 +325,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: b948523af4c873949b458cc0123775cd, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  uiManager: {fileID: 1013115839}
   deploymentPhaseControl: {fileID: 1685173638}
-  cursor: {fileID: 1414521162}
   timeline: {fileID: 1013115834}
 --- !u!1 &230143115
 GameObject:
@@ -545,6 +590,46 @@ RectTransform:
   m_CorrespondingSourceObject: {fileID: 224615319850239198, guid: 78d998e82cc8bd54c8a4cb3e138833f5,
     type: 2}
   m_PrefabInternal: {fileID: 1952889034}
+--- !u!1 &800986979
+GameObject:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 0}
+  serializedVersion: 6
+  m_Component:
+  - component: {fileID: 800986980}
+  - component: {fileID: 800986981}
+  m_Layer: 0
+  m_Name: PiecePlaceControl
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!4 &800986980
+Transform:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 0}
+  m_GameObject: {fileID: 800986979}
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_LocalPosition: {x: 0, y: 0, z: 0}
+  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_Children: []
+  m_Father: {fileID: 959419239}
+  m_RootOrder: 1
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!114 &800986981
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 0}
+  m_GameObject: {fileID: 800986979}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 19067f93f3c3a1c4f81cd0864f7792bc, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!1 &921232818
 GameObject:
   m_ObjectHideFlags: 0
@@ -643,6 +728,7 @@ Transform:
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children:
   - {fileID: 1414521164}
+  - {fileID: 800986980}
   - {fileID: 1685173640}
   - {fileID: 125722435}
   m_Father: {fileID: 0}
@@ -661,7 +747,7 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   cursor: {fileID: 1414521161}
   deployPhaseControl: {fileID: 1685173638}
-  directionalInput: {fileID: 125722436}
+  directionalInput: {fileID: 0}
 --- !u!1 &1013115833
 GameObject:
   m_ObjectHideFlags: 0
@@ -1049,6 +1135,9 @@ MonoBehaviour:
   board: {fileID: 597238866}
   uiManager: {fileID: 1013115839}
   startCoords: {x: 0, y: 0}
+  currCoords: {x: 0, y: 0}
+  selectedUnit: {fileID: 0}
+  selectedUnitOwner: {fileID: 0}
   movementEnabled: 1
 --- !u!212 &1414521163
 SpriteRenderer:
@@ -1150,7 +1239,7 @@ Transform:
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children: []
   m_Father: {fileID: 959419239}
-  m_RootOrder: 1
+  m_RootOrder: 2
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1 &1865771652
 GameObject:
