+++ Assets/Scripts/UI/Control.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4f9fb0e87bc37fd409b52304085aab81
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
